Version 1.7.4 of the AWS Java SDK
This release updates the Elastic Load Balancing client with support for connection draining and updates the Amazon CloudFront client with support for HTTPS redirects and Server Name Indication (SNI) as well as several smaller enhancements and bug fixes for the SDK.

+++ src/main/java/com/amazonaws/services/s3/internal/crypto/AdjustedRangeInputStream.java
@@ -68,7 +68,7 @@ public class AdjustedRangeInputStream extends InputStream {
         }
         // The number of bytes the user may read is equal to the number of the bytes in the range.
         // Note that the range includes the endpoints.
-        this.virtualAvailable = (int)(rangeEnd - rangeBeginning) + 1;
+        this.virtualAvailable = (rangeEnd - rangeBeginning) + 1;
     }
 
     /* (non-Javadoc)
